Refactor data structures
