+++ src/views/GalleryView.vue
@@ -1,23 +1,22 @@
 <script setup lang="ts">
-import { ARTWORKS_NEWEST_FIRST } from '@/data/artworks';
-import ThumbnailCard from '@/components/ThumbnailCard.vue';
+import { ARTWORKS } from '@/data/artworks'
+import ThumbnailCard from '@/components/ThumbnailCard.vue'
+import { computed } from 'vue'
+import { reverse_chronological, type Thumbnail } from '@/core/Thumbnail'
+
+const artworks_newest_first = computed<Thumbnail[]>(() => {
+  return ARTWORKS.map((x) => x.thumbnail)
+    .filter((x) => !x.hide)
+    .sort(reverse_chronological)
+})
 </script>
 
 <template>
-    <h1 class="centered">Gallery</h1>
-    <div class="tableau">
-        <template v-for="artwork in ARTWORKS_NEWEST_FIRST" :key="artwork.sort_key">
-            <ThumbnailCard 
-                v-if="artwork.show"
-                :image_url="artwork.thumbnail_url"
-                :alt="artwork.alt_text"
-                :link="artwork.artwork_url"
-                :title="artwork.title"
-                :years="artwork.date"
-            />
-        </template>
-        <div class="plaque" v-show="true">
-            🚧 Curating old artworks, check back later for more!
-        </div>
-    </div>
+  <h1 class="centered">Gallery</h1>
+  <div class="tableau">
+    <template v-for="thumbnail in artworks_newest_first" :key="thumbnail.sort_key">
+      <ThumbnailCard :card="thumbnail" />
+    </template>
+    <div class="plaque" v-show="true">🚧 Curating old artworks, check back later for more!</div>
+  </div>
 </template>
